feat(test): refactor all PolicyDefinitionStore implementors tests by using a base class
