+++ extensions/control-plane/store/cosmos/policy-store-cosmos/src/main/java/org/eclipse/dataspaceconnector/cosmos/policy/store/CosmosPolicyStoreExtension.java
@@ -17,6 +17,7 @@ package org.eclipse.dataspaceconnector.cosmos.policy.store;
 import dev.failsafe.RetryPolicy;
 import org.eclipse.dataspaceconnector.azure.cosmos.CosmosClientProvider;
 import org.eclipse.dataspaceconnector.azure.cosmos.CosmosDbApiImpl;
+import org.eclipse.dataspaceconnector.spi.monitor.Monitor;
 import org.eclipse.dataspaceconnector.spi.policy.store.PolicyDefinitionStore;
 import org.eclipse.dataspaceconnector.spi.security.Vault;
 import org.eclipse.dataspaceconnector.spi.system.Inject;
@@ -24,13 +25,19 @@ import org.eclipse.dataspaceconnector.spi.system.Provides;
 import org.eclipse.dataspaceconnector.spi.system.ServiceExtension;
 import org.eclipse.dataspaceconnector.spi.system.ServiceExtensionContext;
 import org.eclipse.dataspaceconnector.spi.system.health.HealthCheckService;
+import org.eclipse.dataspaceconnector.spi.types.TypeManager;
 
 @Provides({ PolicyDefinitionStore.class })
 public class CosmosPolicyStoreExtension implements ServiceExtension {
 
     @Inject
-    private RetryPolicy retryPolicy;
+    private RetryPolicy<Object> retryPolicy;
 
+    @Inject
+    private TypeManager typeManager;
+
+    @Inject
+    private Monitor monitor;
     @Inject
     private CosmosClientProvider clientProvider;
 
@@ -47,7 +54,7 @@ public class CosmosPolicyStoreExtension implements ServiceExtension {
 
         var cosmosDbApi = new CosmosDbApiImpl(configuration, clientProvider.createClient(vault, configuration));
 
-        var store = new CosmosPolicyDefinitionStore(cosmosDbApi, context.getTypeManager(), retryPolicy, configuration.getPartitionKey());
+        var store = new CosmosPolicyDefinitionStore(cosmosDbApi, typeManager, retryPolicy, configuration.getPartitionKey(), monitor);
         context.registerService(PolicyDefinitionStore.class, store);
 
         context.getTypeManager().registerTypes(PolicyDocument.class);
