+++ extensions/control-plane/store/sql/policy-store-sql/build.gradle.kts
@@ -31,7 +31,7 @@ dependencies {
     testImplementation(project(":extensions:common:junit"))
     testImplementation(testFixtures(project(":core:common:util")))
     testImplementation("org.postgresql:postgresql:${postgresVersion}")
-
+    testImplementation(testFixtures(project(":spi:control-plane:policy-spi")))
 }
 
 publishing {
