+++ extensions/control-plane/store/sql/policy-store-sql/src/main/java/org/eclipse/dataspaceconnector/sql/policy/store/schema/postgres/PolicyDefinitionMapping.java
@@ -25,6 +25,8 @@ import org.eclipse.dataspaceconnector.sql.translation.TranslationMapping;
 public class PolicyDefinitionMapping extends TranslationMapping {
     public PolicyDefinitionMapping(SqlPolicyStoreStatements statements) {
         add("uid", statements.getPolicyIdColumn());
+        add("id", statements.getPolicyIdColumn());
+        add("createdAt", statements.getCreatedAtColumn());
         add("policy", new PolicyMapping(statements));
     }
 }
